requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
chromadb==0.5.20
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1